requirements.txt
Django==2.1.2
django-cors-headers==2.4.0
django-admin-rangefilter==0.3.12
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-framework-multi-slug-field==0.2.0
django-notifications-hq==1.5.0
djangorestframework==3.9.0
djangorestframework-datatables==0.4.0
django-crispy-forms==1.7.2
mccabe==0.6.1
mysqlclient==1.3.14
pycodestyle==2.4.0
pydocstyle==3.0.0
pyflakes==2.0.0
pylama==7.6.5
pytz==2018.6
six==1.11.0
snowballstemmer==1.2.1

# django admin theme
django-jet==1.0.8

# config
django-environ==0.4.4

# for django eva2
django-eav2==0.11.1

# import and export 
django-import-export==1.1.0

#auto computations for totals in admin
admin-totals==1.0.1

# pdf generation
WeasyPrint>=0.34
xhtml2pdf==0.2.3